Move test that requires non-static method source, from TestLambdaOps to a new test class
